Add network provider interface and implementations
Add the plumbing for a network provider interface.

Add two implementations of this interface; one for a no-op
and one to support neutron.

Also move _build_client from the Neutron DHCP provider to a
more common location, to allow it to be used for the
neutron_plugin network provider.

This patch also adds a network_provider attribute to nodes.

Change-Id: If9560a7845fa586fe2c3c7b5b2c8ed5e68caafa2
Partial-bug: #1526403
Co-Authored-By: Om Kumar (om.kumar@hp.com)